@DMwangnima 2024-06-06
# Key Characteristic
Random Key with length of 30.
Generate Func:
```golang
func getRandomString(length int) string {
	var resBuilder strings.Builder
	resBuilder.Grow(length)
	corpus := "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
	rand.Seed(time.Now().UnixNano() + int64(100))
	for i := 0; i < length; i++ {
		resBuilder.WriteByte(corpus[rand.Intn(len(corpus))])
	}
	return resBuilder.String()
}
```
# Benchmark Result:
- New version (without Cache)
```
BenchmarkConsistPicker_RandomDistributionKey
BenchmarkConsistPicker_RandomDistributionKey/10ins
BenchmarkConsistPicker_RandomDistributionKey/10ins-12         	 2417481	       508.9 ns/op	      48 B/op	       1 allocs/op
BenchmarkConsistPicker_RandomDistributionKey/100ins
BenchmarkConsistPicker_RandomDistributionKey/100ins-12        	 2140726	       534.6 ns/op	      48 B/op	       1 allocs/op
BenchmarkConsistPicker_RandomDistributionKey/1000ins
BenchmarkConsistPicker_RandomDistributionKey/1000ins-12       	 2848216	       407.7 ns/op	      48 B/op	       1 allocs/op
BenchmarkConsistPicker_RandomDistributionKey/10000ins
BenchmarkConsistPicker_RandomDistributionKey/10000ins-12      	 2701766	       492.7 ns/op	      48 B/op	       1 allocs/op
```
- Old version (with cache)
```
BenchmarkConsistPicker_RandomDistributionKey
BenchmarkConsistPicker_RandomDistributionKey/10ins
BenchmarkConsistPicker_RandomDistributionKey/10ins-12         	  671974	      1959 ns/op	     363 B/op	       9 allocs/op
BenchmarkConsistPicker_RandomDistributionKey/100ins
BenchmarkConsistPicker_RandomDistributionKey/100ins-12        	  568735	      2182 ns/op	     379 B/op	       9 allocs/op
BenchmarkConsistPicker_RandomDistributionKey/1000ins
BenchmarkConsistPicker_RandomDistributionKey/1000ins-12       	  576940	      2358 ns/op	     486 B/op	       9 allocs/op
BenchmarkConsistPicker_RandomDistributionKey/10000ins
BenchmarkConsistPicker_RandomDistributionKey/10000ins-12      	  475820	      2702 ns/op	     515 B/op	       9 allocs/op
```